inference_server/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
pillow
python-multipart